feat(components/molecule/stepper): add steps state to the provider
